rearrange dto directory location
